Move `unsort`/`unorder` to tests
