+++ src/MultipleTesting.jl
@@ -16,8 +16,6 @@ import Distributions: estimate
 
 import SpecialFunctions: digamma
 
-import Random: shuffle!
-
 export PValues,
     adjust,
     PValueAdjustment,
